Add SettingsProfiles, an additional ledger-CLI recognized date format, and extract an "environment".
SettingsProfiles are a new proto and set of flags that allow you to
create and save named runtime configurations. This allows an invocation
to use less flags overall and instead lean on known working
configurations.

Ledger-CLI files can now specify a date using slashes instead of just
dashes.

Add JcfEnvironment which is a better abstraction and encapsulation
behind the runtime flags of CsvMatcher. I would like to eventually
refactor the TUI to also use this same environment.